Refactor KTable to control loading visibility with useLoading composable
